shaderpipeline: More work: caps, refactoring, legacy, etc. etc.

+++ panda/src/display/graphicsStateGuardian.h
@@ -226,7 +226,7 @@ PUBLISHED:
   INLINE void set_shader_model(ShaderModel shader_model);
   MAKE_PROPERTY(shader_model, get_shader_model, set_shader_model);
 
-  INLINE int get_supported_shader_capabilities() const;
+  INLINE uint64_t get_supported_shader_capabilities() const;
   MAKE_PROPERTY(supported_shader_capabilities, get_supported_shader_capabilities);
 
   virtual int get_supported_geom_rendering() const;
@@ -638,7 +638,7 @@ protected:
 
   ShaderModel _auto_detect_shader_model;
   ShaderModel _shader_model;
-  int _supported_shader_caps;
+  uint64_t _supported_shader_caps;
 
   static PT(TextureStage) _alpha_scale_texture_stage;
